analytics/requirements.txt
apscheduler==3.8.1
connexion==2.10.0
PyYAML < 6
requests==2.26.0
SQLAlchemy==1.4.27
pymongo
pymysql[srv]
dnspython